Iterate over existing video frames instead of extracting video segments
For V3.0 datasets, we would use ffmpeg to extract video segments corresponding to episodes.
This had some issues:
  - If the segments had to be re-encoded, it would be super slow
  - Sometimes copy codec wouldn't work (not sure of all the failure modes)

Instead of creating temporary video files corresponding to episode segments, this new approach simply just iterates over particular regions of the original dataset video corresponding to episode segments.
It is pretty fast and robust as it will seek to the i-frame nearest to the start time and iterate through the rest of the duration of the video segment.
This helps avoid encoding related issues with creating temporary episode segment video files to operate on.

Topic: reddyshr/main/iterate_frames_in_range